Rename common.ts to types.ts
[review feedback]
